Clean Code Changes Phase 3

+++ Functions/src/main/java/com/epam/engx/cleancode/functions/task2/Account.java
@@ -23,7 +23,7 @@ public abstract class Account implements User {
     
     private int sumOfReviewAnswers(){
     	
-    	return(getAllReviews().stream().map(r -> r.getAnswers().size()).reduce(0,(sum,reviewAnswers) -> sum + reviewAnswers));
+    	return (int) (getAllReviews().stream().map(r -> r.getAnswers().size()).reduce(0,(sum,reviewAnswers) -> sum + reviewAnswers));
 		/*
 		 * int reviewAnswers = 0; for (Review r : getAllReviews()) reviewAnswers +=
 		 * r.getAnswers().size(); return reviewAnswers
